Parse enum objects and values (mostly), start writing parsing logic for StructObjects

+++ Serialization/Data/EnumData.cs
@@ -9,10 +9,10 @@ namespace NotInfiltrator.Serialization.Data
 {
     public class EnumData : Data
     {
-        public UInt16 NameId { get; init; } = 0;
-        public UInt32 ObjReference { get; init; } = 0;
+        public Int16 NameId { get; init; } = 0;
+        public Int32 ObjReference { get; init; } = 0;
 
-        public string Name => StructBin.GetString(NameId);
+        public string Name => StructBin.GetString((ushort)NameId);
 
         public EnumData(int id, StructBin sbin) : base(id, sbin) { }
     }
